Handle submission of existing storage and network maps
